refactor: 이름들을 자동차 객체로 변환하는 로직을 RacingCars로 이동
